+++ app/models/user.rb
@@ -2328,6 +2328,11 @@ TRIGGER
     name.present? ? name : username
   end
 
+  def purge_redis_vizjson_cache
+    redis_keys = CartoDB::Visualization::Collection.new.fetch(user_id: self.id).map(&:redis_vizjson_key)
+    CartoDB::Visualization::Member.redis_cache.del redis_keys unless redis_keys.empty?
+  end
+
   private
 
   def name_exists_in_organizations?